Use notifier to check job status and remove listener
