+++ src/main/java/org/jenkinsci/plugins/lightstepincidentresponse/LightstepWebhookPublisher.java
@@ -3,6 +3,7 @@
  * This source code is licensed under the MIT license found in the LICENSE file in the root directory of this source tree.
  */
 package org.jenkinsci.plugins.lightstepincidentresponse;
+import hudson.model.Result;
 import hudson.tasks.Notifier;
 import hudson.Launcher;
 import hudson.model.AbstractBuild;
@@ -16,9 +17,22 @@ import hudson.util.ListBoxModel;
 import java.io.IOException;
 import java.util.LinkedHashMap;
 import java.util.Map;
+import java.util.logging.Logger;
+import java.lang.*;
+import jenkins.model.Jenkins;
+import okhttp3.MediaType;
+import okhttp3.OkHttpClient;
+import okhttp3.Request;
+import okhttp3.RequestBody;
+import okhttp3.Response;
+import org.json.simple.JSONValue;
+import org.json.simple.JSONObject;
 import org.kohsuke.stapler.DataBoundConstructor;
 
-public class LightstepWebhookPublisher extends Notifier{
+public class LightstepWebhookPublisher extends Notifier {
+
+	private static final MediaType JSON_MEDIA_TYPE = MediaType.parse("application/json; charset=utf-8");
+	private static final Logger log = Logger.getLogger(LightstepWebhookPublisher.class.getName());
 
 	public String webHookUrl;
 	public Boolean onFailure;
@@ -52,7 +66,108 @@ public class LightstepWebhookPublisher extends Notifier{
 	@Override
 	public boolean perform(AbstractBuild<?, ?> build, Launcher launcher, BuildListener listener)
 			throws InterruptedException, IOException {
+
+		Result result = build.getResult();
+		String res = result.toString();
+
+		if (result == null) {
+			log.severe("No build result.");
+		}
+		String webHookUrl = this.webHookUrl;
+		if (webHookUrl.isEmpty()) {
+			log.severe("No webhook URL provided.");
+		}
+		String buildUrl = "";
+		if (Jenkins.get().getRootUrl() == null) {
+			buildUrl = build.getUrl();
+		} else {
+			buildUrl = Jenkins.get().getRootUrl() + build.getUrl();
+		}
+		String jobName = build.getProject().getDisplayName();
+		String buildName = build.getDisplayName();
+		Integer buildNum = build.number;
+		String buildId = build.getId().toString();
+		String buildStatusURL = build.getBuildStatusUrl().toString();
+		String buildStatusSummary = build.getBuildStatusSummary().toString();
+		String buildDuration = build.getDurationString();
+		String buildTime = build.getTime().toString();
+
+		JSONObject event = new JSONObject();
+		event.put("buildName", buildName);
+		event.put("jobName", jobName);
+		event.put("buildUrl", buildUrl);
+		event.put("buildNum", buildNum);
+		event.put("buildId", buildId);
+		event.put("buildStatusURL", buildStatusURL);
+		event.put("buildStatusSummary", buildStatusSummary);
+		event.put("buildDuration", buildDuration);
+		event.put("buildTime", buildTime);
+		event.put("source_url", buildUrl);
+
+		if (this.selectedStatus) {
+			String severity = "";
+			String status = "";
+			switch (res) {
+				case "SUCCESS":
+					if (this.onResolve) {
+						severity = "clear";
+						status = "resolved";
+					}
+					break;
+				case "FAILURE":
+					if (this.onFailure) {
+						severity = this.failureSeverity;
+						status = "failure";
+					}
+					break;
+				case "UNSTABLE":
+					if (this.onUnstable) {
+						severity = this.unstableSeverity;
+						status = "unstable";
+					}
+					break;
+				case "ABORTED":
+					if (this.onAborted) {
+						severity = this.abortedSeverity;
+						status = "aborted";
+					}
+					break;
+			}
+			if (!severity.isEmpty() && !status.isEmpty()) {
+				event.put("status", status);
+				event.put("severity", severity);
+				log.info("Build Payload " + event);
+				httpPost(webHookUrl, event);
+			}
+		}
 		return true;
+	}
+
+	private void httpPost(String url, Object object) {
+		try {
+			String jsonString = JSONValue.toJSONString(object);
+			RequestBody body = RequestBody.create(jsonString, JSON_MEDIA_TYPE);
+			Request request = new Request.Builder().url(url).post(body).build();
+			OkHttpClient client = new OkHttpClient();
+			Response response = client.newCall(request).execute();
+			String responseBody = response.body().string();
+
+			try {
+				if (response.code() == 200) {
+					log.info("Webhook invocation successful " + responseBody);
+				} else {
+					log.severe("Webhook invocation failed " + responseBody);
+				}
+			} catch (Exception e) {
+				log.severe("Exception occurred " + url + e);
+			} finally {
+				if (response != null) {
+					response.close();
+				}
+			}
+		} catch (Exception e) {
+			log.severe("Exception " + url + e);
+		}
 	}
 
 	@Override
@@ -67,7 +182,7 @@ public class LightstepWebhookPublisher extends Notifier{
 		public boolean isApplicable(Class<? extends AbstractProject> jobType) {
 			return true;
 		}
-
+	
 		@Override
 		public String getDisplayName() {
 			return "Send alerts to Lightstep Incident Response";
